requirements.txt
mfrc522>=1.4.3
RPi.GPIO>=0.7.1
requests>=2.28.0 